Last thing we need is image similarity. FYI- The whole team is here, not just Drew. My computer is the only one that works...

+++ src/templates/deleteDB.py
@@ -1,9 +1,9 @@
 from pymongo import MongoClient
 
-def deleteAll():
+def deleteAll(username):
     my_client = MongoClient("mongodb+srv://team_andrew:Green@cluster1.jsqyd.mongodb.net/test")
     db = my_client.ImageSearch #connect to "ImageSearch" Database
     collection = db.get_collection("ImageData") #connect to "ImageData" Collection
-    x = collection.delete_many({})
+    x = collection.delete_many({"user_id": username})
 
     print(x.deleted_count, "documents deleted")
